refactor: extract module FileUtil
